Add a SECRET_KEY_SOURCE environment variable
It was too tricky to get this app up and running, but with a
`SECRET_KEY_SOURCE` environment variable it's easier. Setting this
variable will import the secret key into the newly-created keyring.

After import the environment variable is cleared. This way the data is
not leaked.

Set this variable in the shell where you run `rails s`, but there's no
need to set it in the shell where you run `rails c`. Specifically, it
only needs to be done once per `PGPPATH`.