Passphrases
Add a passphrase option as well. The passphrase is encoded using SHA256
before being sent over the wire, in an effort to reduce the need to
trust us and to make it harder to get the user's actual passphrase.

Depends on the DOMCrypt extension, which is a W3C draft.